+++ Net_Core/Framework/Models/UI/UIGroupProfile.cs
@@ -4,8 +4,16 @@ public class UIGroupProfile
 {
     public UIGroupProfile()
     {
+        Id = -1;
         RolesInGroup = new string[]{};
         RolesNotInGroup = new string[]{};
+    }
+
+    public UIGroupProfile(MGroupProfile groupProfile)
+    {
+        Description = groupProfile.Description;
+        Id = groupProfile.Id;
+        Name = groupProfile.Name;
     }
 
     public int Id { get; set; }