+++ Net_Core/Web.Support/Utilities/FeedbackUtility.cs
@@ -1,7 +1,10 @@
+using System;
 using System.Data;
+using System.Text.RegularExpressions;
 using GrowthWare.BusinessLogic;
 using GrowthWare.Framework;
 using GrowthWare.Framework.Models;
+using GrowthWare.Framework.Models.Base;
 
 namespace GrowthWare.Web.Support.Utilities;
 
@@ -23,6 +26,7 @@ public static class FeedbackUtility
 
     public static UIFeedback GetFeedback(int feedbackId)
     {
+        testProfile();
         UIFeedback mRetVal = getBusinessLogic.GetFeedback(feedbackId);
         return mRetVal;
     }
@@ -32,5 +36,28 @@ public static class FeedbackUtility
         UIFeedback mRetVal = null;
         mRetVal = getBusinessLogic.SaveFeedback(feedback);
         return mRetVal;
+    }
+
+    private static void testProfile()
+    {
+        bool mUseBrackets = true;
+        string mPrimaryKeyName = "GroupSeqId";
+        MGroupProfile mProfile = GroupUtility.GetGroupProfile(1);
+
+        string mTableName = mProfile.TableName;
+        
+        string mDeleteWithParameters = MGroupProfile.GenerateDeleteWithParameters(mPrimaryKeyName, mUseBrackets);
+        string mDeleteWithValues = mProfile.GenerateDeleteWithValues<MGroupProfile>(mPrimaryKeyName, mUseBrackets);
+        string mDeleteWithValuesSpecifyKeyValue = MGroupProfile.GenerateDeleteWithValues(mPrimaryKeyName, "'yaba'", mUseBrackets);
+        string mInsertWithParameters = MGroupProfile.GenerateInsertWithParameters<MGroupProfile>(mUseBrackets);
+        string mInsertWithValues = mProfile.GenerateInsertWithValues<MGroupProfile>(mUseBrackets);
+        string mUpdateWithParameters = MGroupProfile.GenerateUpdateWithParameters<MGroupProfile>(mPrimaryKeyName, mUseBrackets);
+        string mUpdateWithValues = mProfile.GenerateUpdateWithValues<MGroupProfile>(mPrimaryKeyName, mUseBrackets);
+        string mUpdateWithValuesSpecifyKeyValue = mProfile.GenerateUpdateWithValues<MGroupProfile>(mPrimaryKeyName, "'yaba'", mUseBrackets);
+
+        DataTable mDataTable = MGroupProfile.GenerateEmptyTable<MGroupProfile>("Accounts", false);
+        mDataTable = null;
+        mDataTable = MGroupProfile.GenerateEmptyTable<MGroupProfile>("Accounts", true);
+        string mStop = string.Empty;
     }
 }